Fix an error
I know this is a read-only version of this package.  However, this method did not pass one of my simple tests, when NA exists and V is nonsingular. This is due to this error in the code. Of course, the fix can be more sophisticated and I just wanted to point to it.